Promjenjen tip podatka iz Long u Double za latitude i longitude

+++ api/src/main/java/hr/fer/hydro/stations/dto/StationRequestDto.java
@@ -8,9 +8,9 @@ public record StationRequestDto(
         String stationName,
 
         @NonNull
-        Long longitude,
+        Double longitude,
 
         @NonNull
-        Long latitude
+        Double latitude
 ) {
 }
